Add files via upload
This set of files is the 3rd update to my set of backgrounds to replace the files of gbz35 theme. They are primarily aimed at small screens, ideally 640x480 such as the piboy dmg, or the lcl gba with its slightly lower profile. The set currently covers around fifty machines, along with additional files to fill in gaps such as emulators and menu logos. instructions are in the rar file. All the files have been named 'background.png' so they are a simple copy and paste. Equally the folders are default names. I created extras to cover some of the alternates, such as neogeo  cd that can be (amongst others) neogeo_cd or neocdz. There are US and Euro files in there too for Super Famicom/SNES. Some of the logo files also have alternates, such as the Odyssey/Videopac. They are simple hand drawn images, made to suit the way I wanted my handheld to look. Every background has the year the machine was released (not an easy task, too many cross overs and country specific releases). Its taken a lot of work to get this far, so forgive any rough edges. There are still a few left to do, and they are being done, but the sets near complete so here it is. V3 is my most complete set. If anyone uses them and has a machine that needs doing to fill the gap, let me know and I'll bump it up the list of remaining images.